RavenDB-25080 Moved the test that allocates over 2GB into StressTests/X64.
