+++ test/SlowTests/Corax/GrowableBitArrayTests.cs
@@ -54,21 +54,7 @@ public class GrowableBitArrayTests : NoDisposalNeeded
             Assert.True(lookup.Add(i), i.ToString());
         }
     }
-
-    [RavenFact(RavenTestCategory.Corax)]
-    public void CanSetSecondBitmap()
-    {
-        using var allocator = new ByteStringContext(SharedMultipleUseFlag.None);
-        using var lookup = new GrowableBitArray(allocator, GrowableBitArray.MaxCapacityPerBitmapInBits + 2);
-        Assert.True(lookup.Add(GrowableBitArray.MaxCapacityPerBitmapInBits));
-        Assert.False(lookup.Add(GrowableBitArray.MaxCapacityPerBitmapInBits));
-        Assert.True(lookup.Add(GrowableBitArray.MaxCapacityPerBitmapInBits+1));
-        Assert.False(lookup.Add(GrowableBitArray.MaxCapacityPerBitmapInBits+1));
-        Assert.True(lookup.Add(GrowableBitArray.MaxCapacityPerBitmapInBits+2));
-        Assert.False(lookup.Add(GrowableBitArray.MaxCapacityPerBitmapInBits+2));
-        Assert.Throws<ArgumentOutOfRangeException>(() => lookup.Add(GrowableBitArray.MaxCapacityPerBitmapInBits+3));
-    }
-
+    
     [RavenTheory(RavenTestCategory.Corax)]
     [InlineDataWithRandomSeed]
     [InlineDataWithRandomSeed]
